+++ packages/common/src/hooks/useUrlQuery/index.ts
@@ -1,1 +1,2 @@
 export * from './useUrlQuery';
+export * from './useHandleQueryParams';
