+++ packages/host/src/routers/DistributionRouter.tsx
@@ -34,6 +34,8 @@ export const DistributionRouter: FC = () => {
   const gotoCustomerRequisition = useMatch(fullCustomerRequisitionPath);
   const gotoCustomers = useMatch(fullCustomersPath);
 
+  console.log("ROUTER")
+
   if (gotoOutboundShipment) {
     return <InvoiceService />;
   }
