+++ packages/invoices/src/OutboundShipment/ListView/ListView.tsx
@@ -12,8 +12,7 @@ import {
   NothingHere,
   useToggle,
   createQueryParamsStore,
-  useUrlQuery,
-  Column,
+  useHandleQueryParams,
 } from '@openmsupply-client/common';
 import { getStatusTranslator, isOutboundDisabled } from '../../utils';
 import { Toolbar } from './Toolbar';
@@ -37,22 +36,14 @@ export const OutboundShipmentListViewComponent: FC = () => {
 
   const { data, isError, isLoading, sort, pagination, filter } =
     useOutbound.document.list();
-  const { urlQuery, updateQuery } = useUrlQuery();
-  const { onChangeSortBy, sortBy } = sort;
+  const { urlQuery, updateSortQuery, onChangeUrlQuery } =
+    useHandleQueryParams(sort);
+  const { sortBy } = sort;
   useDisableOutboundRows(data?.nodes);
 
   useEffect(() => {
-    const sortKey = urlQuery?.sort ?? 'otherPartyName';
-    const column = columns.find(col => col.key === sortKey);
-    console.log('column', column);
-    if (column?.key !== sortBy.key)
-      onChangeSortBy(column as Column<OutboundRowFragment>);
+    onChangeUrlQuery(columns);
   }, [urlQuery]);
-
-  const updateURL = (column: any) => {
-    const sort = column.key;
-    updateQuery({ sort });
-  };
 
   const columns = useColumns<OutboundRowFragment>(
     [
@@ -84,11 +75,9 @@ export const OutboundShipmentListViewComponent: FC = () => {
       ],
       'selection',
     ],
-    { onChangeSortBy: updateURL, sortBy },
+    { onChangeSortBy: updateSortQuery, sortBy },
     [sortBy]
   );
-
-  console.log('columns', columns);
 
   return (
     <>
